+++ packages/core-js/internals/numeric-range-iterator.js
@@ -7,7 +7,7 @@ var isObject = require('../internals/is-object');
 var defineProperties = require('../internals/object-define-properties').f;
 var DESCRIPTORS = require('../internals/descriptors');
 
-var INCORRECT_RANGE = 'Incorrect Number.range arguments';
+var INCORRECT_RANGE = 'Incorrect Iterator.range arguments';
 var NUMERIC_RANGE_ITERATOR = 'NumericRangeIterator';
 
 var setInternalState = InternalStateModule.set;
@@ -17,6 +17,7 @@ var $RangeError = RangeError;
 var $TypeError = TypeError;
 
 var $RangeIterator = createIteratorConstructor(function NumericRangeIterator(start, end, option, type, zero, one) {
+  // TODO: Drop the first `typeof` check after removing lagacy methods in `core-js@4`
   if (typeof start != type || (end !== Infinity && end !== -Infinity && typeof end != type)) {
     throw $TypeError(INCORRECT_RANGE);
   }