+++ packages/core-js/modules/esnext.bigint.range.js
@@ -5,6 +5,7 @@ var NumericRangeIterator = require('../internals/numeric-range-iterator');
 
 // `BigInt.range` method
 // https://github.com/tc39/proposal-Number.range
+// TODO: Remove from `core-js@4`
 if (typeof BigInt == 'function') {
   $({ target: 'BigInt', stat: true, forced: true }, {
     range: function range(start, end, option) {