+++ packages/core-js/modules/esnext.number.range.js
@@ -4,6 +4,7 @@ var NumericRangeIterator = require('../internals/numeric-range-iterator');
 
 // `Number.range` method
 // https://github.com/tc39/proposal-Number.range
+// TODO: Remove from `core-js@4`
 $({ target: 'Number', stat: true, forced: true }, {
   range: function range(start, end, option) {
     return new NumericRangeIterator(start, end, option, 'number', 0, 1);