+++ packages/core-js/proposals/iterator-range.js
@@ -1,0 +1,3 @@
+// https://github.com/tc39/proposal-Number.range
+require('../modules/esnext.iterator.constructor');
+require('../modules/esnext.iterator.range');