+++ packages/core-js/stage/1.js
@@ -5,11 +5,11 @@ require('../proposals/array-last');
 require('../proposals/array-unique');
 require('../proposals/collection-methods');
 require('../proposals/collection-of-from');
+require('../proposals/iterator-range');
 require('../proposals/keys-composition');
 require('../proposals/math-extensions');
 require('../proposals/math-signbit');
 require('../proposals/number-from-string');
-require('../proposals/number-range');
 require('../proposals/object-iteration');
 require('../proposals/observable');
 require('../proposals/pattern-matching');
@@ -20,6 +20,7 @@ require('../proposals/string-cooked');
 // TODO: Obsolete versions, remove from `core-js@4`:
 require('../proposals/array-from-async');
 require('../proposals/map-upsert');
+require('../proposals/number-range');
 require('../proposals/string-replace-all');
 
 module.exports = parent;